% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_capture}
\title{HTML dependencies used by capture}
\usage{
html_dependency_capture(type = c("image", "pdf", "lookbook"))
}
\arguments{
\item{type}{Type of dependency to use.}
}
\value{
an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
HTML dependencies used by capture
}
