% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loading}
\alias{loading}
\title{Loading indicator to be displayed when generating screenshot}
\usage{
loading(
  text = "Generating PDF, please wait...",
  type = c("standard", "hourglass", "circle", "arrows", "dots", "pulse"),
  color = "#246abe",
  background = "rgba(0,0,0,0.8)",
  size = "80px",
  ...
)
}
\arguments{
\item{text}{Text to be displayed below loading animation.}

\item{type}{Type of loading animation.}

\item{color}{Color for text and loading indicator.}

\item{background}{Background color.}

\item{size}{Size (in pixels).}

\item{...}{Other arguments.}
}
\value{
a \code{list} that can be used in \code{\link[=capture]{capture()}} or \code{\link[=capture_pdf]{capture_pdf()}}.
}
\description{
Loading indicator to be displayed when generating screenshot
}
