% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{Capture Screenshot as Image}
\usage{
capture(
  selector,
  filename,
  ...,
  format = c("png", "jpeg"),
  scale = NULL,
  inputId = NULL,
  options = NULL
)
}
\arguments{
\item{selector}{A CSS selector, for example \code{body} to target the whole page or \verb{#<ID>} to target a specific output.}

\item{filename}{Name of the file (without extension) that will be created. If \code{NULL} no file is downloaded client side.}

\item{...}{Arguments passed to HTML button.}

\item{format}{Format of output between: \code{"png"} or \code{"jpeg"}.}

\item{scale}{Scale factor applied to image's dimension. Can be used to get a higher resolution image.}

\item{inputId}{An \code{inputId} to retrieve image as base64 in an \code{input} slot server-side.}

\item{options}{Options (as a list) passed to \href{https://github.com/tsayen/dom-to-image}{domtoimage}
method, for example you can use \code{bgcolor} to set background color.}
}
\value{
an HTML tag.
}
\description{
Add a button to take a screenshot of a specified element and download a PNG file.
}
\note{
It's only possible to take screenshot of elements that are actually visible on screen. It doesn't work in Internet Explorer.
}
\examples{
library(shiny)
library(capture)

ui <- fluidPage(
  tags$h2("Capture example"),
  capture(
    selector = "body",
    filename = "all-page",
    icon("camera"), "Take screenshot of all page",
    format = "jpeg"
  ),
  tags$br(),
  fluidRow(
    column(
      width = 4,
      wellPanel(
        tags$b("Parameters :"),

        selectInput(
          inputId = "loi",
          label = "Law:",
          choices = c("normal", "uniform", "exponential")
        )
      )
    ),
    column(
      width = 8,
      tags$div(
        id = "result-block",
        tags$b("Results :"),
        plotOutput(outputId = "plot"),
        uiOutput(outputId = "mean"),
        verbatimTextOutput(outputId = "raw")
      ),
      capture(
        selector = "#result-block",
        filename = "results-screenshot",
        icon("camera"), "Take screenshot of results",
        options = list(backgroundColor = "#FFF")
      ),
      capture(
        selector = "#result-block",
        filename = "results-screenshot",
        icon("camera"), "Take screenshot of results (bigger scale)",
        scale = 3,
        options = list(backgroundColor = "#FFF")
      ),
      capture(
        selector = "#result-block",
        filename = NULL, # no download client side
        icon("camera"), "Take screenshot of results (retrieve server side)",
        inputId = "screenshot",
        options = list(backgroundColor = "#FFF")
      ),
      uiOutput("out")
    )
  )
)

server <- function(input, output, session) {
  
  output$out <- renderUI({
    # # Here we display image back in interface, 
    # # but you can also write image on disk
    # write_png <- function(x, filename) {
    #   x <- sub(".*,", "", x)
    #   x <- base64enc::base64decode(x)
    #   png::writePNG(png::readPNG(x), filename)
    # }
    # write_png(input$screenshot, "myimage.png")
    
    tags$img(src = input$screenshot)
  })
  
  
  distrib_r <- reactive({
    switch(
      input$loi,
      "normal" = rnorm(1000),
      "uniform" = runif(1000),
      "exponential" = rexp(1000)
    )
  })

  output$plot <- renderPlot({
    hist(distrib_r())
  })

  output$mean <- renderUI({
    tags$p(tags$b("The mean is :"), round(mean(distrib_r()), 2))
  })

  output$raw <- renderPrint({
    summary(distrib_r())
  })
}

if (interactive())
  shinyApp(ui, server)
}
