% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_domtoimage}
\title{HTML dependencies used by capture}
\usage{
html_dependency_domtoimage()
}
\value{
an \code{\link[htmltools]{htmlDependency}}.
}
\description{
HTML dependencies used by capture
}
